StencilDef: Fix signed & unsigned int comparison
The _iterBetSwaps variable is a signed integer, since its default value
is -1. This means we need to cast the unsigned integer to a signed
integer, to prevent compilation warnings when building programs which
use the Stencil skeleton.

+++ include/SkelCL/detail/StencilDef.h
@@ -370,7 +370,7 @@ unsigned int Stencil<Tout(Tin)>::determineIterationsBetweenDataSwaps(
     const Matrix<Tin>& /*in*/, unsigned int iterLeft)
 {
   // User choses a value
-  if (iterLeft <= _iterBetSwaps)
+  if (static_cast<int>(iterLeft) <= _iterBetSwaps)
     return iterLeft;
   else if (_iterBetSwaps != -1)
     return _iterBetSwaps;
